Add GPU backend switching
A global variable stores a symbol that is either :CPU or :GPU.  Whichever backend is active dictates which version of training programs to run.  A wrapper function for checkNumGrad() is created in FCANN.jl that always calls the proper version.

The GPU backend will only be initialized if the system has CUDAdrv.jl and CUBLAS.jl installed.  Testing functions have been added for the GPU backend and will ony run on a computer with those packages installed.

File names will change for saving training results to indicate which backend was active during training.  The GPU functions only have  the absErr cost function currently.