Clean up utilities
Adds default lambda to checkNumGrad so it can be called without an argument.  Fixes GPU device initialization with modulus of worker ID.  This ensures a different device will be used for each worker assigned to a node as long as the workers are added contiguously and match the number of devices.