Fix GPU error calculation
Changed absolute value sum to a regular sum.  The absolute value sum was a relic of assuming the absolute error was being used and the activation layer stored the difference.  Now the final layer stores the actual per example error values so for error functions that have negative values, the absolute sum will not work.