Revise calcOutput and add calcMultiOut
Added a function for both CPU and GPU backends to predict output values using an ensemble of parameters.  This implementation sets up required intermediate memory objects once and reuses them for multiple network parameters collecting outputs.  This function is also used when calcMultiOutput() is called which performs all of the ensemble output and error calculations.

calcOutput() and multiOut() now also output two types of error in the case a log likelihood cost function is used.  In this case, all functions that save files containing error values will have two values of train and test error with the specific error function for each named in file headers.