Add indexed output cost function and output
All methods of cost function derivatives are of forms that make use of all of the outputs and combine them with a sum of some kind.  These changes implement a type of output derivative which is just the output at a particular index.  If the output size is already of dimension 1, then this can be used to calculate the identity output of the function without any output data for comparison.  It is another form of output free derivative that is not an autoencoder and where the output data and error calculation is replaced by simply selecting the output value for a single index.