Add index and cross entropy output to GPU
Adds GPU kernels for computing the gradient with respect to an output index and the cross entropy loss of the output softmax with respect to a correct label.  Both of these GPU additions currently only work for single vector examples with an integer index.  For consistency changed most kernel launches to 1D with strided loops in the kernels.  Also added a new calculator for the 1D kernel launch params based on the number of streaming multiprocessors for the user GPU